feat: portfolio view

+++ fractional_share/holdingstock/models.py
@@ -1,12 +1,11 @@
 from django.db import models
-
 # Create your models here.
 class HoldingStock(models.Model):
     portfolio_id = models.ForeignKey('portfolio.Portfolio',on_delete=models.CASCADE,verbose_name="포트폴리오명")
     stock_name=models.CharField(max_length=128, verbose_name='stock_name')
-    stock_share = models.IntegerField(verbose_name='stock_share')
+    stock_share = models.FloatField(verbose_name='stock_share')
     invest_amount = models.IntegerField(verbose_name='invest_amount')
-    earn_rate = models.IntegerField(verbose_name='earn_rate')
+    earn_rate = models.FloatField(verbose_name='earn_rate')
 
 
     class Meta:
