+++ fractional_share/portfolio/models.py
@@ -1,11 +1,10 @@
 from django.db import models
 
-
 # Create your models here.
 class Portfolio(models.Model):
     user_id = models.ForeignKey('member.Member',on_delete=models.CASCADE,verbose_name="포트폴리오명")
     total_invest = models.IntegerField(verbose_name='total_invest')
-    total_rate = models.IntegerField(verbose_name='total_rate')
+    total_rate = models.FloatField(verbose_name='total_rate')
     total_earn = models.IntegerField(verbose_name='total_earn')
